+++ inst/NEWS.md
@@ -1,4 +1,5 @@
 # MetaboDynamics 1.1.6
+- get_ORA_annotations function (retrieving KEGG IDs hierarchies) added back to package
 - no more samples in estimates dynamics as probability of clustering solution will be implemented with a bubbletree in cluster_dynamics
 - estimates now returns a list with: estimated metabolite abundance (mu), 
   estimated standard deviation of metabolite abundance (sigma),